pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "bazi_engine"
version = "0.2.0"
description = "Astronomical BaZi (Four Pillars) calculation engine"
requires-python = ">=3.10"
dependencies = [
    "pyswisseph>=2.10.3",
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0"]
skyfield = ["skyfield>=1.45"]
